Make offset and target_bytes size_t so they don't go negative

+++ artdaq/TransferPlugins/TCPSocketTransfer.cc
@@ -410,9 +410,9 @@ int artdaq::TCPSocketTransfer::receiveFragmentData(RawDataType* destination, siz
 	uint8_t* buff;
 	size_t byte_cnt = 0;
 	int sts;
-	int offset = 0;
+	size_t offset = 0;
 	SocketState state = SocketState::Metadata;
-	int target_bytes = sizeof(MessHead);
+	size_t target_bytes = sizeof(MessHead);
 
 	pollfd pollfd_s;
 	pollfd_s.events = POLLIN | POLLPRI | POLLERR;
